+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+- Add support for `CustomField::Select` and `CustomField::MultiSelect`
+
 ## v0.3.1 - 2020-11-13
 
 - Update metadata in the `teamtailor.gemspec`